// Angular-Playground/platform/projects/rxjs/src/lib/models/people.ts
import { StarWarsType } from "./responseWrapper";

export interface People {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
  type: StarWarsType;
}