// Angular-Playground/platform/projects/rxjs/src/lib/models/planets.ts
import { StarWarsType } from './responseWrapper';

export interface Planets {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
  type: StarWarsType;
}